requirements.txt
langchain-groq
pdfminer.six
langchain
langchain-community
fastapi
uvicorn[standard]
python-dotenv
langchain-pinecone
pinecone-client
langchain-google-genai
langchain-pinecone
pinecone
python-multipart
PyPDF2
google-generativeai